+++ runtime/onert/core/include/backend/basic/MemoryManager.h
