+++ runtime/onert/core/src/backend/basic/MemoryManager.cc
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include <backend/cpu_common/MemoryManager.h>
+#include <backend/basic/MemoryManager.h>
 
 #include <cassert>
